Changed constructor and added image reconstruction
Changed network constructor, so now it will dinamically select the input and output number of neurons needed adjusted based on the input and output vectors of frames. Also added a function where you can train a network to construct images from given inputs, and after training the output of the network can be printed as a picture.